chore: Cleanup header files and remove minishell.h
