+++ include/clean.h
@@ -13,6 +13,8 @@
 #ifndef CLEAN_H
 # define CLEAN_H
 
+# include "defines.h"
+
 void	ft_clean_shell(t_shell *shell);
 void	ft_clean_and_exit_shell(t_shell *shell, int exit_code, char *msg);
 void	free_env_node(t_env *env);
