+++ include/executor.h
@@ -29,8 +29,6 @@ void	handle_control_op(t_shell *shell, t_list_d **cmd_table_node);
 void	handle_assignment(t_shell *shell, t_final_cmd_table *final_cmd_table);
 void	handle_external_cmd(t_shell *shell, t_cmd_table *cmd_table);
 void	handle_builtin(t_shell *shell, t_list_d **cmd_table_node);
-// void	handle_builtin(t_shell *shell,
-// 			t_list_d **cmd_table_node, t_final_cmd_table *final_cmd_table);
 
 /* Redirection */
 bool	restore_std_io(int saved_std_io[2]);
@@ -47,7 +45,6 @@ void	safe_close_pipe(t_pipe *pipe);
 void	handle_pipes_parent(t_pipe *new_pipe, t_pipe *old_pipe);
 void	handle_pipes_child(t_pipe *new_pipe, t_pipe *old_pipe);
 void	safe_close_all_pipes(t_shell *shell);
-// void	safe_move_nonempty_pipe(t_pipe *from, t_pipe *to);
 void	replace_fd(int *from_end, int *to_end);
 
 #endif
