+++ source/debug/show_list.c
@@ -10,8 +10,36 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
 #include "parser.h"
+
+char	*ft_get_token_type_str(int type)
+{
+	if (type == T_WORD)
+		return ("WORD");
+	else if (type == T_ASSIGNMENT_WORD)
+		return ("ASSIGNMENT_WORD");
+	else if (type == T_RED_IN)
+		return ("RED_IN");
+	else if (type == T_RED_OUT)
+		return ("RED_OUT");
+	else if (type == T_HERE_DOC)
+		return ("HEREDOC");
+	else if (type == T_APPEND)
+		return ("APPEND");
+	else if (type == T_PIPE)
+		return ("PIPE");
+	else if (type == T_OR)
+		return ("OR");
+	else if (type == T_AND)
+		return ("AND");
+	else if (type == T_L_BRACKET)
+		return ("L_BRACKET");
+	else if (type == T_R_BRACKET)
+		return ("R_BRACKET");
+	else if (type == T_END)
+		return ("END");
+	return (NULL);
+}
 
 void	print_token(t_token *token)
 {
