+++ source/shell_clean.c
@@ -10,7 +10,6 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
 #include "utils.h"
 #include "executor.h"
